Verifies database adapter on a platform basis
It's normal to have in your Gemfile more than one platform (JRuby, MRI
etc). Railbrazer, though, is not considering it when choosing the
database adapter. It's also raising an exception on Gemfiles such as
the following:

```
gem 'pg', platform: :mri
gem 'activerecord-jdbcpostgresql-adapter', platform: :jruby
```

With this change, we only select the gems that are defined for the
current platform.